requirements.txt
Python == 3.12.2
Flask==3.1.0
flask-cors==6.0.0
python-dotenv==1.0.1
langchain==0.3.20
langchain-community==0.3.19
langchain-core==0.3.41
langchain-groq==0.2.4
groq==0.18.0
